make it better

+++ assets/php/delete_teachers.php
@@ -1,11 +1,13 @@
 <?php
 include 'config.php';
 session_start();
-$id = $_GET['id'];
-$sql = "DELETE FROM Teachers WHERE ID='$id'";
-$result = $db ->query($sql);
-$_SESSION['status'] = "success";
-$_SESSION['message'] = "Teacher Successfully Deleted";
-header("Location: /epm_teachers.php");
-exit();
+if (isset($_GET['id'])){
+    $id = $_GET['id'];
+    $sql = "DELETE FROM Teachers WHERE ID='$id'";
+    $result = $db ->query($sql);
+    $_SESSION['status'] = "success";
+    $_SESSION['message'] = "Teacher Successfully Deleted";
+    header("Location: /epm_teachers.php");
+    exit();
+}
 ?>
